# Refactoring dyn Column