Further refactoring... 

This removes the `DynamicFastFieldReader` and we now just deal with a `Arc<dyn Column>`.
The code ends up simpliied a notch.

I do not observe any regression (nor perf improvment)_.

# After

```
cargo +nightly bench --features unstable bench_intfastfield
    Finished bench [optimized] target(s) in 0.08s
     Running unittests src/lib.rs (target/release/deps/tantivy-66e9d65f044be08d)

running 7 tests
test fastfield::bench::bench_intfastfield_jumpy_fflookup                                                                 ... bench:     450,360 ns/iter (+/- 1,397)
test fastfield::bench::bench_intfastfield_jumpy_veclookup                                                                ... bench:     670,624 ns/iter (+/- 6,486)
test fastfield::bench::bench_intfastfield_scan_all_fflookup                                                              ... bench:     174,085 ns/iter (+/- 2,212)
test fastfield::bench::bench_intfastfield_scan_all_fflookup_gcd                                                          ... bench:     199,005 ns/iter (+/- 3,396)
test fastfield::bench::bench_intfastfield_scan_all_vec                                                                   ... bench:       9,651 ns/iter (+/- 300)
test fastfield::bench::bench_intfastfield_stride7_fflookup                                                               ... bench:      21,976 ns/iter (+/- 243)
test fastfield::bench::bench_intfastfield_stride7_vec                                                                    ... bench:       8,493 ns/iter (+/- 127)
```

# Before
```
❯ cargo +nightly bench --features unstable bench_intfastfield
    Finished bench [optimized] target(s) in 0.07s
     Running unittests src/lib.rs (target/release/deps/tantivy-66e9d65f044be08d)

running 7 tests
test fastfield::bench::bench_intfastfield_jumpy_fflookup                                                                 ... bench:     450,952 ns/iter (+/- 2,179)
test fastfield::bench::bench_intfastfield_jumpy_veclookup                                                                ... bench:     678,692 ns/iter (+/- 9,008)
test fastfield::bench::bench_intfastfield_scan_all_fflookup                                                              ... bench:     174,656 ns/iter (+/- 1,399)
test fastfield::bench::bench_intfastfield_scan_all_fflookup_gcd                                                          ... bench:     199,451 ns/iter (+/- 2,433)
test fastfield::bench::bench_intfastfield_scan_all_vec                                                                   ... bench:       9,671 ns/iter (+/- 92)
test fastfield::bench::bench_intfastfield_stride7_fflookup                                                               ... bench:      21,999 ns/iter (+/- 151)
test fastfield::bench::bench_intfastfield_stride7_vec                                                                    ... bench:       9,103 ns/iter (+/- 1,051)

```